+++ jdecode.go
@@ -1,3 +1,4 @@
+// Library to decode Juniper secret hashes ($9$)
 package junosdecode
 
 import (