+++ CHANGELOG.md
@@ -7,6 +7,8 @@ ENHANCEMENTS:
 * export constant `MagicPrefix`
 * export variable `ErrDiffGapDec`
 * generate an error when missing or extra character(s) detected in encoded string
+* add an example for [godoc](https://pkg.go.dev/github.com/jeremmfr/junosdecode#example_)
+
 BUG FIXES:
 
 * fix conversion of the character in string from the rune number (compatibility with latest golang version)